Cleaned up auth file structure
