+++ app/src/main/java/com/project/freshtomatoes/ui/factories/AuthLoginViewModelFactory.kt
@@ -3,7 +3,7 @@ package com.project.freshtomatoes.ui.factories
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
 import com.project.freshtomatoes.ui.FreshTomatoes
-import com.project.freshtomatoes.ui.pages.EmailPassword.AuthLoginViewModel
+import com.project.freshtomatoes.ui.pages.AuthLogin.AuthLoginViewModel
 
 class AuthLoginViewModelFactory : ViewModelProvider.Factory {
     override fun <T : ViewModel> create(modelClass: Class<T>): T {
