+++ app/src/main/java/com/project/freshtomatoes/ui/factories/AuthSignUpModelFactory.kt
@@ -3,7 +3,7 @@ package com.project.freshtomatoes.ui.factories
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
 import com.project.freshtomatoes.ui.FreshTomatoes
-import com.project.freshtomatoes.ui.pages.EmailPassword.AuthSignUpViewModel
+import com.project.freshtomatoes.ui.pages.AuthSignUp.AuthSignUpViewModel
 
 class AuthSignUpModelFactory: ViewModelProvider.Factory {
     override fun <T : ViewModel> create(modelClass: Class<T>): T {
