+++ app/src/main/java/com/project/freshtomatoes/ui/pages/AuthLogin/AuthLoginScreen.kt
@@ -1,11 +1,9 @@
-package com.project.freshtomatoes.ui.pages.EmailPassword
+package com.project.freshtomatoes.ui.pages.AuthLogin
 
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.rememberScrollState
 import androidx.compose.foundation.verticalScroll
@@ -15,35 +13,28 @@ import androidx.compose.material3.Text
 import androidx.compose.material3.TextField
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.collectAsState
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.saveable.rememberSaveable
-import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import androidx.lifecycle.viewmodel.compose.viewModel
 import com.project.freshtomatoes.LocalNavController
+import com.project.freshtomatoes.data.AuthState
 import com.project.freshtomatoes.ui.Router
 import com.project.freshtomatoes.ui.components.AuthStatus
 import com.project.freshtomatoes.ui.components.PasswordField
-import com.project.freshtomatoes.ui.factories.AuthSignUpModelFactory
+import com.project.freshtomatoes.ui.factories.AuthLoginViewModelFactory
 
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
-fun AuthSignUpScreen(viewmodel: AuthSignUpViewModel = viewModel(factory = (AuthSignUpModelFactory()))) {
+fun AuthLoginScreen(viewmodel: AuthLoginViewModel = viewModel(factory = AuthLoginViewModelFactory())) {
     val navController = LocalNavController.current
 
     val email = viewmodel.email.collectAsState()
     val password = viewmodel.password.collectAsState()
-    var confirmPassword = viewmodel.confirmPassword.collectAsState()
     val errorMessage = viewmodel.errorMessage.collectAsState()
-    val signupState = viewmodel.signupState.collectAsState()
-
-    var passwordVisibility = rememberSaveable { mutableStateOf(false) }
+    val loginState = viewmodel.loginState.collectAsState()
 
     Column(
         modifier = Modifier
@@ -52,30 +43,33 @@ fun AuthSignUpScreen(viewmodel: AuthSignUpViewModel = viewModel(factory = (AuthS
             .verticalScroll(rememberScrollState()),
         horizontalAlignment = Alignment.CenterHorizontally
     ) {
-        Text("Sign Up", fontSize = 32.sp, fontWeight = FontWeight.Bold, modifier = Modifier.padding(20.dp))
-        TextField(value = email.value, onValueChange = { viewmodel.setEmail(it) }, label = { Text("Email") }, modifier = Modifier.padding(20.dp))
-        PasswordField("Password", password, { viewmodel.setPassword(it) }, passwordVisibility)
-        PasswordField("Confirm Password", confirmPassword, { viewmodel.setConfirmPassword(it) }, passwordVisibility)
+        Text("Log In", fontSize = 32.sp, fontWeight = FontWeight.Bold, modifier = Modifier.padding(20.dp))
 
-        Spacer(modifier = Modifier.height(10.dp))
+        TextField(
+            value = email.value,
+            onValueChange = { viewmodel.setEmail(it) },
+            label = { Text("Email") },
+            modifier = Modifier.padding(20.dp)
+        )
+        PasswordField("Password", password, { viewmodel.setPassword(it) })
 
         Row(horizontalArrangement = Arrangement.SpaceEvenly, modifier = Modifier.fillMaxWidth()) {
-            Button(onClick = { viewmodel.signUp() } ) {
-                Text("Sign Up")
+            Button(onClick = { navController.navigate(Router.SignUp.route) }) {
+                Text("Sign up")
             }
-            Button(onClick = { navController.navigate(Router.Account.route) }) {
+            Button(onClick = { viewmodel.signIn() }) {
                 Text("Log In")
             }
         }
         Row(horizontalArrangement = Arrangement.SpaceEvenly, modifier = Modifier.fillMaxWidth()) {
-            Text("                           ")
-            Text("Already have an account?")
+            Text("Create an Account     ")
+            Text("                      ")
         }
 
-        AuthStatus(signupState, errorMessage, "Processing signup...")
+        AuthStatus(loginState, errorMessage, "Processing login...")
     }
 
-    if(signupState.value == AuthState.Success) {
+    if(loginState.value == AuthState.Success) {
         navController.navigate(Router.Home.route)
     }
 }
