+++ app/src/main/java/com/project/freshtomatoes/ui/pages/AuthSignUp/AuthSignUpScreen.kt
@@ -1,9 +1,11 @@
-package com.project.freshtomatoes.ui.pages.EmailPassword
+package com.project.freshtomatoes.ui.pages.AuthSignUp
 
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.rememberScrollState
 import androidx.compose.foundation.verticalScroll
@@ -13,28 +15,33 @@ import androidx.compose.material3.Text
 import androidx.compose.material3.TextField
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.collectAsState
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.saveable.rememberSaveable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import androidx.lifecycle.viewmodel.compose.viewModel
 import com.project.freshtomatoes.LocalNavController
+import com.project.freshtomatoes.data.AuthState
 import com.project.freshtomatoes.ui.Router
 import com.project.freshtomatoes.ui.components.AuthStatus
 import com.project.freshtomatoes.ui.components.PasswordField
-import com.project.freshtomatoes.ui.factories.AuthLoginViewModelFactory
+import com.project.freshtomatoes.ui.factories.AuthSignUpModelFactory
 
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
-fun AuthLoginScreen(viewmodel: AuthLoginViewModel = viewModel(factory = AuthLoginViewModelFactory())) {
+fun AuthSignUpScreen(viewmodel: AuthSignUpViewModel = viewModel(factory = (AuthSignUpModelFactory()))) {
     val navController = LocalNavController.current
 
     val email = viewmodel.email.collectAsState()
     val password = viewmodel.password.collectAsState()
+    var confirmPassword = viewmodel.confirmPassword.collectAsState()
     val errorMessage = viewmodel.errorMessage.collectAsState()
-    val loginState = viewmodel.loginState.collectAsState()
+    val signupState = viewmodel.signupState.collectAsState()
+
+    var passwordVisibility = rememberSaveable { mutableStateOf(false) }
 
     Column(
         modifier = Modifier
@@ -43,33 +50,30 @@ fun AuthLoginScreen(viewmodel: AuthLoginViewModel = viewModel(factory = AuthLogi
             .verticalScroll(rememberScrollState()),
         horizontalAlignment = Alignment.CenterHorizontally
     ) {
-        Text("Log In", fontSize = 32.sp, fontWeight = FontWeight.Bold, modifier = Modifier.padding(20.dp))
+        Text("Sign Up", fontSize = 32.sp, fontWeight = FontWeight.Bold, modifier = Modifier.padding(20.dp))
+        TextField(value = email.value, onValueChange = { viewmodel.setEmail(it) }, label = { Text("Email") }, modifier = Modifier.padding(20.dp))
+        PasswordField("Password", password, { viewmodel.setPassword(it) }, passwordVisibility)
+        PasswordField("Confirm Password", confirmPassword, { viewmodel.setConfirmPassword(it) }, passwordVisibility)
 
-        TextField(
-            value = email.value,
-            onValueChange = { viewmodel.setEmail(it) },
-            label = { Text("Email") },
-            modifier = Modifier.padding(20.dp)
-        )
-        PasswordField("Password", password, { viewmodel.setPassword(it) })
+        Spacer(modifier = Modifier.height(10.dp))
 
         Row(horizontalArrangement = Arrangement.SpaceEvenly, modifier = Modifier.fillMaxWidth()) {
-            Button(onClick = { navController.navigate(Router.SignUp.route) }) {
-                Text("Sign up")
+            Button(onClick = { viewmodel.signUp() } ) {
+                Text("Sign Up")
             }
-            Button(onClick = { viewmodel.signIn() }) {
+            Button(onClick = { navController.navigate(Router.Account.route) }) {
                 Text("Log In")
             }
         }
         Row(horizontalArrangement = Arrangement.SpaceEvenly, modifier = Modifier.fillMaxWidth()) {
-            Text("Create an Account     ")
-            Text("                      ")
+            Text("                           ")
+            Text("Already have an account?")
         }
 
-        AuthStatus(loginState, errorMessage, "Processing login...")
+        AuthStatus(signupState, errorMessage, "Processing signup...")
     }
 
-    if(loginState.value == AuthState.Success) {
+    if(signupState.value == AuthState.Success) {
         navController.navigate(Router.Home.route)
     }
 }
